feat(eval): golden Q&A matching in core -- the producer for GOLDEN_EXPECTED_ANSWER
#378 (U4) gave the SDK the consumer side of answer-key grounding
(per_session_context + CategoricalContextSource.GOLDEN_EXPECTED_ANSWER);
the producer stayed in scripts/quality_report.py. This extracts it
verbatim into bigquery_agent_analytics.golden_matching:

- embed_texts(texts, model=None, batch_size=50, max_attempts=5) --
  L2-normalised vectors with bounded 429/503 retry (the exact resilience
  added after an embedding quota burst aborted a live gate run in #385);
- DEFAULT_GOLDEN_THRESHOLD = 0.92;
- match_golden_qa(question_by_key, golden_qa, threshold) -> per-key
  judge-context mapping (expected answers + out-of-scope decline notes)
  ready for evaluate_categorical(per_session_context=...), plus per-key
  match metadata for reporting. Keys are opaque: exact
  ResolvedTraceSelectors on the BigQuery path, session ids locally.

scripts/quality_report.py becomes a thin caller (aliases keep its
public/test surface stable); the embedding-retry tests move to
tests/test_golden_matching.py alongside new matching tests (threshold,
out-of-scope decline note, empty inputs, opaque keys).

Follow-up direction (tracked in the promotion issue): a server-side
embedding path via ai_ml_integration's AI.EMBED would remove the genai
embedding dependency entirely.
