+++ src/bigquery_agent_analytics/__init__.py
@@ -390,6 +390,25 @@ except ImportError as e:
       e,
   )
 
+# Golden Q&A matching (producer for GOLDEN_EXPECTED_ANSWER judge context)
+try:
+  from .golden_matching import DEFAULT_GOLDEN_THRESHOLD
+  from .golden_matching import embed_texts
+  from .golden_matching import match_golden_qa
+
+  __all__.extend(
+      [
+          "DEFAULT_GOLDEN_THRESHOLD",
+          "embed_texts",
+          "match_golden_qa",
+      ]
+  )
+except ImportError as e:
+  logger.debug(
+      "Could not import golden matching components: %s.",
+      e,
+  )
+
 # Categorical Views
 try:
   from .categorical_views import CategoricalViewManager
